Some fixes to W4 seminar code.
Fixed memleak in renderthread on adding new commands. When we have 100
of entities which add commands each frame it really is noticable.
Fix for leak in constantly adding queries in flecs (need to research if
this is intended behaviour or a bug in framework).
Fix for release build libraries not being copied _before_ build